Rename cmd/lefty/io.h to cmd/lefty/leftyio.h
to avoid conflicts with Windows io.h
