+++ cmd/lefty/lefty.c
@@ -23,7 +23,7 @@
 #include "common.h"
 #include "g.h"
 #include "code.h"
-#include "io.h"
+#include "leftyio.h"
 #include "mem.h"
 #include "tbl.h"
 #include "parse.h"
